Fixed Issues Added User wise data

+++ src/main/kotlin/com/triveous/todo/app/repositories/TodoRepoImpl.kt
@@ -16,18 +16,18 @@ import kotlin.reflect.jvm.internal.impl.load.kotlin.JvmType
 class TodoRepoImpl(@Autowired private val jdbcTemplate: JdbcTemplate) : TodoRepository {
 
     companion object {
-        const val SQL_CREATE = "INSERT INTO TRVS_TODO(ID,NAME,REMINDER_TIME,STATUS,PRIORITY) values (NEXTVAL('TRVS_TODO_SEQ'),?,?,?,?)"
-        const val SQL_GET_ALL_TODOS = "SELECT * FROM TRVS_TODO ORDER BY PRIORITY"
-        const val SQL_DELETE_TODO_WITH_ID = "DELETE FROM TRVS_TODO WHERE ID = ?"
-        const val SQL_UPDATE_TODO = "UPDATE TRVS_TODO SET NAME= ?, REMINDER_TIME= ?,STATUS= ?, PRIORITY = ? WHERE ID= ?"
-        const val SQL_COUNT_BY_NAME = "SELECT COUNT(*) FROM TRVS_TODO WHERE NAME= ?"
-        const val SQL_GET_COMPLETED_TODOS = "SELECT * FROM TRVS_TODO WHERE STATUS = 1"
-        const val SQL_GET_INCOMPLETE_TODOS = "SELECT * FROM TRVS_TODO WHERE STATUS = 0"
-        const val SQL_UPDATE_STATUS = "UPDATE TRVS_TODO SET STATUS = ? WHERE ID = ?"
-        const val SQL_UPDATE_PRIORITY = "UPDATE TRVS_TODO SET PRIORITY = ? WHERE ID = ?"
+        const val SQL_CREATE = "INSERT INTO TRVS_TODO(ID,NAME,REMINDER_TIME,STATUS,PRIORITY,USER_ID) values (NEXTVAL('TRVS_TODO_SEQ'),?,?,?,?,?)"
+        const val SQL_GET_ALL_TODOS = "SELECT * FROM TRVS_TODO WHERE user_id = ? ORDER BY PRIORITY"
+        const val SQL_DELETE_TODO_WITH_ID = "DELETE FROM TRVS_TODO WHERE (ID = ? AND USER_ID = ?)"
+        const val SQL_UPDATE_TODO = "UPDATE TRVS_TODO SET NAME= ?, REMINDER_TIME= ?,STATUS= ?, PRIORITY = ? WHERE ID= ? AND USER_ID = ?"
+        const val SQL_COUNT_BY_NAME = "SELECT COUNT(*) FROM TRVS_TODO WHERE (NAME = ? AND USER_ID =?)"
+        const val SQL_GET_COMPLETED_TODOS = "SELECT * FROM TRVS_TODO WHERE (STATUS = 1 AND  USER_ID = ?)"
+        const val SQL_GET_INCOMPLETE_TODOS = "SELECT * FROM TRVS_TODO WHERE (STATUS = 0 AND  USER_ID = ?)"
+        const val SQL_UPDATE_STATUS = "UPDATE TRVS_TODO SET STATUS = ? WHERE (ID = ? AND  USER_ID = ?)"
+        const val SQL_UPDATE_PRIORITY = "UPDATE TRVS_TODO SET PRIORITY = ? WHERE ( ID = ? AND  USER_ID = ?)"
     }
 
-    override fun createTodo(name: String, reminder_time: Int, status: Int, priority: Int) {
+    override fun createTodo(name: String, reminder_time: Int, status: Int, priority: Int,user_id: Int) {
         val keyHolder: KeyHolder = GeneratedKeyHolder()
         try {
             jdbcTemplate.update({
@@ -36,6 +36,7 @@ class TodoRepoImpl(@Autowired private val jdbcTemplate: JdbcTemplate) : TodoRepo
                 ps.setInt(2, reminder_time)
                 ps.setInt(3, status)
                 ps.setInt(4, priority)
+                ps.setInt(5,user_id)
                 return@update ps
             }, keyHolder)
         } catch (e: Exception) {
@@ -44,45 +45,45 @@ class TodoRepoImpl(@Autowired private val jdbcTemplate: JdbcTemplate) : TodoRepo
 
     }
 
-    override fun getAllTodos(): Collection<Todo> {
-        return jdbcTemplate.query(SQL_GET_ALL_TODOS, todoRowMapper)
+    override fun getAllTodos(user_id: Int): Collection<Todo> {
+        return jdbcTemplate.query(SQL_GET_ALL_TODOS, arrayOf<Any>(user_id), todoRowMapper)
     }
 
-    override fun deleteTodoWithId(id: Int): Int {
-        return jdbcTemplate.update(SQL_DELETE_TODO_WITH_ID, id)
+    override fun deleteTodoWithId(id: Int,user_id: Int): Int {
+        return jdbcTemplate.update(SQL_DELETE_TODO_WITH_ID, id,user_id)
     }
 
-    override fun editTodo(name: String?, status: Int?, priority: Int?, reminder_time: Int?, id: Int) {
+    override fun editTodo(name: String?, status: Int?, priority: Int?, reminder_time: Int?, id: Int,user_id: Int) {
         try {
-            jdbcTemplate.update(SQL_UPDATE_TODO, name, reminder_time, status, priority, id)
+            jdbcTemplate.update(SQL_UPDATE_TODO, name, reminder_time, status, priority, id,user_id)
         } catch (e: Exception) {
             throw  EtAuthException("COULD NOT UPDATE DUE TO ${e.message}")
         }
     }
 
-    override fun getCountByName(name: String): Int {
-        return jdbcTemplate.queryForObject(SQL_COUNT_BY_NAME, arrayOf<Any>(name), Int::class.java)
+    override fun getCountByName(name: String,user_id: Int): Int {
+        return jdbcTemplate.queryForObject(SQL_COUNT_BY_NAME, arrayOf<Any>(name,user_id), Int::class.java)
     }
 
-    override fun getAllCompletedTodos(): Collection<Todo> {
-        return jdbcTemplate.query(SQL_GET_COMPLETED_TODOS, todoRowMapper)
+    override fun getAllCompletedTodos(user_id: Int): Collection<Todo> {
+        return jdbcTemplate.query(SQL_GET_COMPLETED_TODOS,arrayOf<Any>(user_id), todoRowMapper)
     }
 
-    override fun getIncompleteTodos(): Collection<Todo> {
-        return jdbcTemplate.query(SQL_GET_INCOMPLETE_TODOS, todoRowMapper)
+    override fun getIncompleteTodos(user_id: Int): Collection<Todo> {
+        return jdbcTemplate.query(SQL_GET_INCOMPLETE_TODOS,arrayOf<Any>(user_id), todoRowMapper)
     }
 
-    override fun updateStatus(status: Int?, id: Int) {
+    override fun updateStatus(status: Int?, id: Int,user_id: Int) {
         try {
-            jdbcTemplate.update(SQL_UPDATE_STATUS, status, id)
+            jdbcTemplate.update(SQL_UPDATE_STATUS, status, id,user_id)
         } catch (e: java.lang.Exception) {
             throw  EtAuthException("Could Not update due to ${e.message}")
         }
     }
 
-    override fun updatePriority(priority: Int?, id: Int) {
+    override fun updatePriority(priority: Int?, id: Int,user_id: Int) {
         try {
-            jdbcTemplate.update(SQL_UPDATE_PRIORITY, priority, id)
+            jdbcTemplate.update(SQL_UPDATE_PRIORITY, priority, id,user_id)
         } catch (e: java.lang.Exception) {
             throw  EtAuthException("Could Not update due to ${e.message}")
         }
@@ -93,6 +94,8 @@ class TodoRepoImpl(@Autowired private val jdbcTemplate: JdbcTemplate) : TodoRepo
                 rs.getString("NAME"),
                 rs.getString("REMINDER_TIME"),
                 rs.getInt("STATUS"),
-                rs.getInt("PRIORITY"))
+                rs.getInt("PRIORITY"),
+                rs.getInt("USER_ID")
+                )
     }
 }
